aj.w {st4gi}	R Documentation

Adjust values following the method of Westcott

Description

This function adjust the observed values of an experiment planted following the method described by Westcott (1981) with a grid of checks.

Usage

aj.w(
  trait,
  geno,
  ck1,
  ck2,
  row,
  col,
  ncb = 10,
  nrs = NULL,
  method = c("weighted", "flat"),
  ind = TRUE,
  p = 0.5,
  dfr
)

Arguments

trait
The name of the column for the trait to adjust.

geno
The name of the column that identifies the genotypes.

ck1
Name of check 1.

ck2
Name of check 2.

row
The name of the column that identifies the rows.

col
The name of the column that identifies the columns.

ncb
Number of columns between two check columns.

nrs
Number of rows to span the row of the plot.

method
The method to fit the values. See details.

ind
Logical. See details.

p
The proportion of the check values differences used for the adjustment. See details.

dfr
The name of the data frame.

Details

The values of the selected trait are adjusted using some mean of the values of all the checks located on the row of the plot plus the nrs rows at each side of the row of the plot. If method = "flat" the simple mean of the checks is used for the adjustment, if method = "weighted" a weighted mean is used for the adjustmen, where the checks closer to the plot get more weight. If p = 1 then the values are adjusted in the same proportion that the checks vary around the field, for values lower than 1 the values are adjusted based on that proportion over the checks variation, if p = 0 then there is no adjustment. If ind = TRUE, each check is centered around its own mean before the adjustment, if ind = FALSE, then both checks are centered around the overall mean. If method = "flat", ncb = 5, nrs = 1, and ind = FALSE, then it corresponds to the method proposed by Westcott. If not specified, nrs = floor(ncb / 2). If the layout does not correspond with the Westcott method, then the observed values are adjusted with the values of the checks planted nearby (this is in the rectangular region definied by floor(ncb / 2) columns and rows at each side of the plot) and a warning is issued.

Value

It returns the adjusted values.

Author(s)

Raul Eyzaguirre.

References

Westcott, B. (1981). Two methods for early generation yield assessment in winter wheat. In: Proc. of the 4th meeting of the Biometrics in Plant Breeding Section of Eucarpia. INRA Poitier, France, pp 91-95.

Examples

# Create design
dfr <- cr.w(1:1000, "Dag", "Cem", 40, 10)
dfr <- dfr$book
# Create some random data
dfr$y <- rnorm(dim(dfr)[1])
# Run adjustment
aj.w("y", "geno", "Dag", "Cem", "row", "col", dfr = dfr)

[Package st4gi version 3.1.2 Index]