ammi {st4gi}	R Documentation

AMMI or GGE with data at plot level

Description

This function runs AMMI (Gollob, H. R., 1968) or GGE (Yan , W. et al., 2000) with data at plot level.

Usage

ammi(
  trait,
  geno,
  env,
  rep,
  dfr,
  method = c("ammi", "gge"),
  f = 0.5,
  maxp = 0.1
)

Arguments

trait
The name of the column for the trait to analyze.

geno
The name of the column that identifies the genotypes.

env
The name of the column that identifies the environments.

rep
The name of the column that identifies the replications or blocks. A RCBD is assumed.

dfr
The name of the data frame.

method
"ammi" or "gge".

f
Scaling factor, defaults to 0.5.

maxp
Maximum allowed proportion of missing values to estimate, default is 10%.

Details

Significance of PCs are evaluated only with method = "ammi" and if the data are balanced (with missing values estimated if possible).

Value

It returns an object of class ammi with the overall, genotype, environment and interaction means, the interaction effects matrix, the first and second PC values for genotypes and environments, the ANOVA table, and a table with the contribution of each PC. Significance of PCs are included only if method = "ammi" and the data are balanced (with missing values estimated if possible).

Author(s)

Raul Eyzaguirre.

References

Gollob, H. R. (1968). A Statistical Model which combines Features of Factor Analytic and Analysis of Variance Techniques, Psychometrika, Vol 33(1): 73-114.

Yan, W. et al. (2000). Cultivar evaluation and mega-environment investigation based on the GGE biplot, Crop Sci., Vol 40: 597-605.

See Also

svd

Examples

model.ammi <- ammi("y", "geno", "env", "rep", met8x12)
model.ammi
model.gge <- ammi("y", "geno", "env", "rep", met8x12, method = "gge")
model.gge

[Package st4gi version 3.1.2 Index]