ammi.gxe {st4gi}	R Documentation

AMMI or GGE with data from an interaction means matrix

Description

This function runs AMMI (Gollob, H. R., 1968) or GGE (Yan , W. et al., 2000) with data from an interaction means matrix.

Usage

ammi.gxe(
  int.mean,
  trait.name = NULL,
  method = c("ammi", "gge"),
  f = 0.5,
  aov.model = NULL,
  nrep = NULL
)

Arguments

int.mean
GxE means matrix, genotypes in rows, environments in columns.

trait.name
The name of the column for the trait to analyze.

method
"ammi" or "gge".

f
Scaling factor, defaults to 0.5.

aov.model
Analysis of variance model.

nrep
Number of replications.

Details

Significance of PCs are evaluated only with method = "ammi" and if aov.model and nrep are specified.

Value

It returns an object of class ammi with the overall, genotype, environment and interaction means, the interaction effects matrix, the first and second PC values for genotypes and environments, and a table with the contribution of each PC. ANOVA table and significance of PCs are included only if method = "ammi", and aov.model and nrep are specified.

Author(s)

Raul Eyzaguirre.

References

Gollob, H. R. (1968). A Statistical Model which combines Features of Factor Analytic and Analysis of Variance Techniques, Psychometrika, Vol 33(1): 73-114.

Yan, W. et al. (2000). Cultivar evaluation and mega-environment investigation based on the GGE biplot, Crop Sci., Vol 40: 597-605.

See Also

svd

Examples

# Compute GxE means
int.mean <- tapply(met8x12$y, list(met8x12$geno, met8x12$env), mean, na.rm = TRUE)
# Run AMMI with GxE means matrix
model.ammi <- ammi.gxe(int.mean, trait.name = "y")
model.ammi
# Run GGE with GxE means matrix
model.gge <- ammi.gxe(int.mean, trait.name = "y", method = "gge")
model.gge

[Package st4gi version 3.1.2 Index]