aov.lxt {st4gi}	R Documentation

Analysis for a Line x Tester genetic design.

Description

This is a modified version of the "lineXtester" function of package "agricolae" for the analysis of a Lina x Tester genetic design.

Usage

aov.lxt(trait, line, tester, rep, dfr)

Arguments

trait
The name of the column for the trait to analyze.

line
The name of the column that identifies the lines.

tester
The name of the column that identifies the testers.

rep
The name of the column that identifies the replications.

dfr
The name of the data frame.

Details

Data must correspond with a line by tester design and must be balanced. If there are missing values they are not estimated and results can be dubious.

Value

It returns the analysis for the line by tester genetic design.

Author(s)

Raul Eyzaguirre.

Examples

aov.lxt("yield", "line", "tester", "rep", lxt)

[Package st4gi version 3.1.2 Index]