aov.met {st4gi}	R Documentation

ANOVA for MET with a RCBD

Description

Fit an analysis of variance model for a multi environment trial (MET) with a RCBD in each environment.

Usage

aov.met(trait, geno, env, rep, dfr, maxp = 0.1)

Arguments

trait
The name of the column for the trait to analyze.

geno
The name of the column that identifies the genotypes.

env
The name of the column that identifies the environments.

rep
The name of the column that identifies the replications or blocks.

dfr
The name of the data frame.

maxp
Maximum allowed proportion of missing values to estimate, default is 10%.

Details

If data is unbalanced, missing values are estimated up to an specified maximum proportion, 10% by default. Genotypes and environments are considered as fixed factors while the blocks are considered as random and nested into the environments.

Value

It returns the ANOVA table.

Author(s)

Raul Eyzaguirre.

Examples

aov.met("y", "geno", "env", "rep", met8x12)

[Package st4gi version 3.1.2 Index]