aov.rcbd {st4gi}	R Documentation

ANOVA for a RCBD

Description

Fit an analysis of variance model for a RCBD.

Usage

aov.rcbd(trait, geno, rep, dfr, maxp = 0.1)

Arguments

trait
The name of the column for the trait to analyze.

geno
The name of the column that identifies the genotypes.

rep
The name of the column that identifies the replications.

dfr
The name of the data frame.

maxp
Maximum allowed proportion of missing values to estimate, default is 10%.

Details

If data is unbalanced, missing values are estimated up to an specified maximum proportion, 10% by default.

Value

It returns an ANOVA table.

Author(s)

Raul Eyzaguirre.

Examples

# Get a copy with some missing values for trw and run ANOVA
temp <- pjpz09
temp[c(10, 20, 30), "trw"] <- NA
aov.rcbd("trw", "geno", "rep", temp)

[Package st4gi version 3.1.2 Index]