asl {st4gi}	R Documentation

Add non significant lines

Description

This function adds columns to a datra frame with averages to identify non significant pairwise comparisons.

Usage

asl(dfr, geno, aver, lsd, symbol = NULL)

Arguments

dfr
The name of the data frame.

geno
The name of the column that identifies the genotypes.

aver
The name of the column with the averages.

lsd
Least significant value.

symbol
If NULLL it uses etters by default. See details

Details

The data frame is sorted descending and columns are added to identify non singnificant pairwise comparisons. By default it works with letters unless a different symbol is specified in symbol.

Value

It returns a data frame with averages and non significant lines.

Author(s)

Raul Eyzaguirre.

Examples

# Adjust model
model <- aov.rcbd("trw", "geno", "rep", pjpz09)
# LSD values
lsd <- sqrt(model[3, 3]) * qt(.975, model[3, 1])
# Compute means
ests <- docomp('mean', 'trw', 'geno', dfr = pjpz09)
# Add non significant lines
ests <- asl(ests, 'geno', 'trw', lsd)

[Package st4gi version 3.1.2 Index]