bm2ptb {st4gi}	R Documentation

Biomart to potatobase format

Description

Converts biomart Excel files to potatobase formats.

Usage

bm2ptb(file.name, path.out)

Arguments

file.name
The name of the biomart Excel fieldbook.

path.out
Folder name and path in the working directory to save results; output if NULL.

Details

All labels (lower or upper case) listed in function check.names.pt are converted to CO variable numbers.

Value

It returns a design file and a data file ready to upload in potatobase.

Author(s)

Raul Eyzaguirre.

[Package st4gi version 3.1.2 Index]