cdt.sp {st4gi}	R Documentation

Compute derived traits for sweetpotato

Description

Compute derived traits for a given fieldbook data frame.

Usage

cdt.sp(dfr, method = c("none", "ps", "np"), value = NULL, nops = NULL)

Arguments

dfr
The name of the data frame.

method
Method to scale data from plot to hectare level. Options are plot size "ps" and number of plants for a full hectare "np". See details.

value
Value for the method selected in square meters if method = "ps" and in number of plants per hectare if method = "np".

nops
Number of plants sowed per plot.

Details

The data frame must use the labels (lower or upper case) listed in function check.names.sp. Conversion from kilograms per plot to tons per hectare can be done using ps, the plot size, or np, the total number of plants that can be allocated in a full hectare. In both cases computations can be adjusted by the number of harvested plants if available in the fieldbook. For method = "np", nops must be specified to compute non adjusted values.

Value

It returns a data frame with the original and derived traits.

Author(s)

Raul Eyzaguirre.

Examples

cdt.sp(pjpz09)

[Package st4gi version 3.1.2 Index]