check.data.sp {st4gi}	R Documentation

Check consistency for sweetpotato experimental data

Description

Set of rules to check for consistency of sweetpotato experimental data. Data labels must be defined as specified in the Procedures for the evaluation and analysis of sweetpotato trials, ISBN 978-92-9060-522-5.

Usage

check.data.sp(
  dfr,
  f = 5,
  out.mod = c("none", "rcbd", "met"),
  out.max = 4,
  add = NULL,
  print.text = TRUE
)

Arguments

dfr
The name of the data frame.

f
Factor for extreme values detection. See details.

out.mod
Statistical model for outliers' detection. See details.

out.max
Threshold for outliers' detection.

add
Additional quantitative traits.

print.text
Logical, if TRUE the output is printed on screen.

Details

The data frame must use the labels (lower or upper case) listed in function check.names.sp.

Extreme low and high values are detected using the interquartile range. The rule is to detect any value out of the interval [Q_1 - f \times IQR; Q_3 + f \times IQR]. By default f = 5. If f = 0, the detection of extreme values is not executed.

Outliers are detected based on standardized residuals for some statistical models. Options are "rcbd" and "met" for a randomized complete block design and a multienvironment trial with RCBD in each environment. By default the threshold value is out.max = 4.

Value

It returns:

$Inconsist.List, a data.frame with a list of all the rows with some kind of inconsistency.

$Inconsist.Matrix, a data.frame with the positions in the fieldbook data frame where inconsistencies occur. These are coded with: (1) for inconsistencies among traits, (2) for out of range values, (3) for extreme values or outliers.

Author(s)

Raul Eyzaguirre.

Examples

check.data.sp(pjpz09)

[Package st4gi version 3.1.2 Index]