check.freq {st4gi}	R Documentation

Check frequencies for complete replications

Description

Check frequencies for designs with complete replications and one or several environments. This is a wrapper for ck.rcbd and ck.f functions.

Usage

check.freq(trait, geno, env = NULL, rep, dfr)

Arguments

trait
The name of the column for the trait to analyze.

geno
The name of the column that identifies the genotypes.

env
The name of the column that identifies the environments, NULL if there are no environments.

rep
The name of the column that identifies the replications.

dfr
The name of the data frame.

Value

Information about the balance, missing values, and replications of the design.

Author(s)

Raul Eyzaguirre.

Examples

check.freq("trw", "geno", NULL, "rep", pjpz09)
check.freq("rytha", "geno", "env", "rep", megaclones)

[Package st4gi version 3.1.2 Index]