check.names.pt {st4gi}	R Documentation

Check fieldbook traits names for potato

Description

Check that fieldbook traits names correspond with the names defined in https://cropontology.org/term/CO_330:ROOT

Usage

check.names.pt(dfr, add = NULL)

Arguments

dfr
The name of the data frame.

add
Additional traits.

Details

The data frame must use the labels (lower or upper case) listed below. Between parentheses are the CO numbers for variables defined in https://cropontology.org/term/CO_330:ROOT.

—————————– Plot identifiers —————————–

plot : Plot number.

row : Row number position of the plot (row_number is also valid).

col : Column number position of the plot (col_number is also valid).

————————— Classification factors————————–

loc : Locations.

year : Years.

season : Seasons.

env : Environments.

geno : Genotypes (accession_name and instn are also valid).

type : Entry type (clon, check, progeny, parent).

rep : Replications (rep_number is also valid).

block : Blocks (block_number is also valid).

treat : Treatments.

———————————– Plants ———————————

ntp : Number of tubers planted per plot (CO_330:0000265).

npe : Number of plants emerged per plot (CO_330:0000268).

nph : Number of plants harvested per plot (CO_330:0000287).

ppe : Proportion of plants emerged in percentage (CO_330:0000283).

pph : Proportion of plants harvested in percentage (CO_330:0000290).

—————————– Number of tubers —————————–

tntp : Total number of tubers per plot (CO_330:0000304).

tntpl : Total number of tubers per plant (CO_330:0000305).

nmtp : Number of marketable tubers per plot (CO_330:0000293).

nmtpl : Number of marketable tubers per plant (CO_330:0000297).

nnomtp : Number of non-marketable tubers per plot (CO_330:0000300).

nmtci : Number of marketable tubers category I per plot.

nmtcii : Number of marketable tubers category II per plot.

—————————– Weight of tubers —————————–

ttwp : Total tuber weight per plot in kg (CO_330:0000317).

ttwpl : Total tuber weight per plant in kg (CO_330:0000321).

mtwp : Marketable tuber weight per plot in kg (CO_330:0000308).

mtwpl : Marketable tuber weight per plant in kg (CO_330:0000311).

nomtwp : Non-marketable tuber weight per plot in kg (CO_330:0000314).

mtwci : Marketable tuber weight category I per plot in kg.

mtwcii : Marketable tuber weight category II per plot in kg.

ttya : Total tuber yield adjusted by number of plants harvested per plot in tons per ha (CO_330:0000323).

ttyna : Total tuber yield not adjusted in tons per ha (CO_330:0000324).

mtya : Marketable tuber yield adjusted by number of plants harvested per plot in tons per ha (CO_330:0000327).

mtyna : Marketable tuber yield not adjusted in tons per ha (CO_330:0000330).

atw : Average tuber weight in g (CO_330:0000333).

atmw : Average marketable tuber weight in g (CO_330:0000336).

———————— Plant and tuber appearance ————————

plant_unif : Plant uniformity estimating 1-9 (CO_330:0000272).

plant_vigor : Plant vigor estimating 1-9 (CO_330:0000276).

se : Senescence estimating 1-9 (CO_330:0000280).

tuber_apper : Tuber appearance estimating 1-9 (CO_330:0000348).

tub_unif : Tuber uniformity estimating 1-9 (CO_330:0000352).

tub_size : Tuber size estimating 1-9 (CO_330:0000356).

tbskn1 : Predominant tuber skin color (CO_330:0000249).

tbfsh1 : Predominant tuber flesh color (CO_330:0000253).

tbshp1 : Tuber shape (CO_330:0000256).

tbshp3 : Tuber depth eyes (CO_330:0000258).

dorpd : Tuber dormancy period (CO_330:0000397).

——————————- Organoleptic ——————————-

chipping : Chips color 1-5 (CO_330:0000384).

ffr : French fries color 1-5 (CO_330:0000388).

aocp : Chips oil absorption rate (CO_330:0000395).

flavour : Tuber flavor after cooking (CO_330:0000379).

textac : Tuber texture after cooking (CO_330:0000380).

cookqu : Tuber cooking quality (CO_330:0000381).

cootime : Tuber cooking time (CO_330:0000389).

———————————- Biomass ———————————

tafw : Total aerial fresh weight per plot in g.

tbfwp : Total biomass fresh weight per plot in g. (CO_330:0000829)

tbfwpl : Total biomass fresh weight per plant in g (CO_330:0000799).

hi_fw : Harvest index fresh weight in percentage (CO_330:0000738).

tadw : Total aerial dry weight per plot in g.

tbdwp : Total biomass dry weight per plot in g. (CO_330:0000828)

tbdwpl : Total biomass dry weight per plant in g (CO_330:0000798).

hi_dw : Harvest index dry weight in percentage (CO_330:0000735).

——————————– Dry weight ——————————–

fwts : Fresh weight of tuber for dry matter determination.

dwts : Dry weight of tuber for dry matter determination.

fwts1 : Fresh weight of tuber sample 1.

fwts2 : Fresh weight of tuber sample 2.

dwts1 : Dry weight of tuber sample 1.

dwts2 : Dry weight of tuber sample 2.

dm1 : Tuber dry matter content sample 1.

dm2 : Tuber dry matter content sample 2.

dm : Tuber dry matter content in percentage (CO_330:0000390).

————————— Nutrients evaluations ————————–

fedw : Tuber iron concentration on dry weight basis in mg per 100 g (CO_330:0000402).

fefw : Tuber iron concentration on fresh weight basis in mg per 100 g (CO_330:0000407).

fedw_xrf : Tuber iron concentration on dry weight basis in mg per kg by XRF (CO_330:0000404).

zndw : Tuber zinc concentration on dry weight basis in mg per 100 g (CO_330:0000408).

znfw : Tuber zinc concentration on fresh weight basis in mg per 100 g (CO_330:0000412).

zndw_xrf : Tuber zinc concentration on dry weight basis in mg per kg by XRF (CO_330:0000409).

antho_dw : Tuber anthocyanin concentration on dry weight basis in mg per 100 g (CO_330:0000414).

antho_fw : Tuber anthocyanin concentration on fresh weight basis in mg per 100 g (CO_330:0000417).

aah_dw : Tuber hydrophilic antioxidant activity on dry weight basis in ugTroloxequiv/g (CO_330:0000418).

aah_fw : Tuber hydrophilic antioxidant activity on fresh weight basis in ugTroloxequiv/g (CO_330:0000421).

aal_dw : Tuber lipophilic antioxidant activity on dry weight basis in ugTroloxequiv/g (CO_330:0000422).

aal_fw : Tuber lipophilic antioxidant activity on fresh weight basis in ugTroloxequiv/g (CO_330:0000425).

asc_dw : Tuber ascorbic acid concentration on dry weight basis in mg per 100 g (CO_330:0000426).

asc_fw : Tuber ascorbic acid concentration on fresh weight basis in mg per 100 g (CO_330:0000429).

pro : Tuber protein content in percentage (CO_330:0000432).

star : Tuber starch content in percentage (CO_330:0000435).

fruc : Tuber fructose content in percentage (CO_330:0000438).

gluc : Tuber glucose content in percentage (CO_330:0000441).

sucr : Tuber sucrose content in percentage (CO_330:0000444).

malt : Tuber maltose content in percentage (CO_330:0000447).

fiber : Tuber fiber content in percentage (CO_330:0000450).

tgly_fw : Tuber glycoalkaloid concentration on fresh weight basis in mg per 100 g (CO_330:0000661).

———————————- Deseases ——————————–

lb : Late Blight severity estimating percentage (CO_330:0000359).

lbi : Late Blight severity estimating percentage evaluation i (i = 1, 2,... 8).

audpc : Late blight AUDPC computed (CO_330:0000363).

raudpc : Late blight relative AUDPC computed (CO_330:0000367).

saudpc : Late blight susceptibility computed (CO_330:0000371).

pvx : Potato virus X resistance (CO_330:0000373).

pvy : Potato virus Y resistance (CO_330:0000374).

prlv : Potatoleaf roll virus resistance (CO_330:0000375).

bw : Bacterial wilt resistance (CO_330:0000376).

rkn : Root knot nematode resistance (CO_330:0000377).

——————————- Other traits ——————————-

rd : Root density estimating 1-3 (CO_330:0000834).

rl : Root length in cm (CO_330:0000764).

sg : Tuber specific gravity (CO_330:0000394).

dsi : Drought susceptibility index computed (CO_330:0000776).

dti : Drought tolerance index computed (CO_330:0000779).

leaflet_tw : Leaflet turgid weight in mg (CO_330:0000818).

leaflet_twi : Leaflet turgid weight evaluation i (i = 1,..., 5).

pw_ev : Degree of canopy wilting estimating 1-9 (CO_330:0000773).

pw_evi : Degree of canopy wilting estimating 1-9 evaluation i (i = 1,..., 5).

snpp : Stem number per plant (CO_330:0000851).

snppi : Stem number per plant evaluation i (i = 1,..., 5).

nipp : Number of inflorescences per plant estimating 0-7 (CO_330:0000226).

nfwp : Number of flowers per main inflorescence estimating 0-5 (CO_330:0000227).

Num_Stolon : Number of stolons estimating 1-9 (CO_330:0000340).

Leng_Stolon : Length of stolons estimating 1-9 (CO_330:0000344).

plahe_ev : Plant height in cm (CO_330:0000758).

plahe_evi : Plant height evaluation i (i = 1,..., 5).

sd_ev : Stem diameter in mm (CO_330:0000847).

sd_evi : Stem diameter evaluation i (i = 1,..., 5)

chlspad_ev : Chlorophyll content index using SPAD-meter in SPAD units (CO_330:0000805).

chlspad_evi : Chlorophyll content index evaluation i (i = 1,..., 5).

cr_ev : Canopy reflectance using Green-Seeker-meter in NDVI units (CO_330:0000860).

cr_evi : Canopy reflectance evaluation i (i = 1,..., 5).

lfa_ev : Leaflet area using Compu Eye, Leaf Symptom Area software in cm^2 (CO_330:0000766).

lfa_evi : Leaflet area using Compu Eye, Leaf Symptom Area software in cm^2 evaluation i (i = 1,..., 5).

rwc_ev : Relative water content in percentage (CO_330:0000741).

rwc_evi : Relative water content evaluation i (i = 1,..., 5).

sla_ev : Leaflet area using EasyLeafArea software in cm^2 per g (CO_330:0000812).

sla_evi : Leaflet area using EasyLeafArea software in cm^2 per g evaluation i (i = 1,..., 5).

Value

It returns:

The fieldbook data frame with all column names in lowercase and with some possible modifications in the names.

A list of warnings for all the column names that have been changed.

A list of warnings for all the column names not recognized.

Author(s)

Raul Eyzaguirre, Johan Ninanya.

Examples

check.names.pt(potatoyield)

[Package st4gi version 3.1.2 Index]