check.names.sp {st4gi}	R Documentation

Check fieldbook traits names for sweetpotato

Description

Check that fieldbook factors and traits names correspond with the names defined in https://cropontology.org/term/CO_331:ROOT and in the Procedures for the evaluation and analysis of sweetpotato trials, ISBN 978-92-9060-522-5.

Usage

check.names.sp(dfr, add = NULL)

Arguments

dfr
The name of the data frame.

add
Additional traits.

Details

The data frame must use the labels (lower or upper case) listed below. Between parentheses are the CO numbers for variables defined in https://cropontology.org/term/CO_331:ROOT or COMP numbers defined in https://sweetpotatobase.org/.

—————————– Plot identifiers —————————–

plot : Plot number.

row : Row number position of the plot (row_number is also valid).

col : Column number position of the plot (col_number is also valid).

————————— Classification factors————————–

loc : Locations.

year : Years.

season : Seasons.

env : Environments.

geno : Genotypes (accession_name and cipno are also valid).

type : Entry type (clon, check, progeny, parent).

rep : Replications (rep_number is also valid).

block : Blocks (block_number is also valid).

treat : Treatment.

harvest : Harvest time.

—————————— Roots and vines —————————–

nops : Number of plants planted per plot (CO_331:0000678).

nope : Number of plants established per plot (CO_331:0000192).

noph : Number of plants harvested per plot (CO_331:0000679).

vir : Virus symptoms estimating 1-9 (CO_331:0000193).

vir1 : Virus symptoms estimating 1-9, first evaluation, 1 month after planting (COMP:0000023).

vir2 : Virus symptoms estimating 1-9, second evaluation, 1 month before harvest (COMP:0000024).

alt : Alternaria symptoms estimating 1-9 (CO_331:0000198).

alt1 : Alternaria symptoms estimating 1-9, first evaluation, 1 month after planting (COMP:0000026).

alt2 : Alternaria symptoms estimating 1-9, second evaluation, 1 month before harvest (COMP:0000022).

vv : Vine vigor estimating 1-9 (CO_331:0000197).

vw : Weight of vines per plot in kg (CO_331:0000227).

nopr : Number of plants with storage roots per plot (CO_331:0000211).

nocr : Number of commercial storage roots per plot (CO_331:0000214).

nonc : Number of non-commercial storage roots per plot (CO_331:0000217).

crw : Weight of commercial storage roots per plot in kg (CO_331:0000220).

ncrw : Weight of non-commercial storage roots per plot in kg (CO_331:0000223).

scol : Storage root predominant skin color estimating 1-9 (CO_331:0000175).

fcol : Storage root predominant flesh color estimating 1-9 (CO_331:0000178).

fcol2 : Storage root secondary flesh color estimating 0-9 (CO_331:0000179).

fcol.cc : Root flesh color from RHS color chart, 1-30.

rs : Storage root size estimating 1-9 (CO_331:0000184).

rf : Storage root appearance estimating 1-9 (CO_331:0000202).

rtshp : Storage root shape estimating 1-9 (CO_331:0000181).

damr : Storage root damage estimating 1-9 (CO_331:0000206).

rspr : Sprouting ability estimating 1-9 (CO_331:0000277).

alcdam : Alcidodes sp. damage estimating 1-9 (CO_331:0000806).

wed : Reaction to sweet potato weevil estimating 1-9 (CO_331:0000207).

stspwv : Reaction to striped sweet potato weevil estimating 1-9 (CO_331:0000720).

milldam : Millipede damage estimating 1-9 (CO_331:0000805).

————————— Dry matter assesment —————————

dmf : Fresh weight of storage root samples for dry matter assessment measuring g (CO_331:0000243).

dmd : Dry weight of storage root samples for dry matter assessment measuring g (CO_331:0000247).

dm : Storage root dry matter content computing percent (CO_331:0000297).

dmvf : Fresh weight of vines samples for dry matter assessment measuring g (CO_331:0000251).

dmvd : Dry weight of vines samples for dry matter assessment measuring g (CO_331:0000255).

dmv : Vine dry matter content computing percent (CO_331:0001014).

———————- Raw and cooked roots attributes ———————

fraw : Fiber content in raw storage roots estimating 1-9 (CO_331:0001038).

suraw : Content of total sugars in raw storage roots estimating 1-9 (CO_331:0000766).

straw : Content of starch in raw storage roots estimating 1-9 (CO_331:0001055).

coof : Fiber content in boiled storage roots estimating 1-9 (CO_331:0000259).

coosu : Storage root sweetness in cooked samples estimating 1-9 (CO_331:0000261).

coost : Storage root texture in boiled samples estimating 1-9 (CO_331:0000265).

coot : Boiled storage root flesh taste estimating 1-9 (CO_331:0000269).

cooap : Boiled storage root overall appearance estimating 1-9 (CO_331:0000273).

————————— Nutrients evaluations ————————–

prot : Content of protein in dry weight basis in raw storage roots measuring percentage (CO_331:0000278).

fe : Content of iron in dry weight basis measuring mg per 100 g (CO_331:0001016).

zn : Content of zinc in dry weight basis measuring mg per 100 g (CO_331:0001017).

ca : Content of calcium in dry weight basis measuring mg per 100 g (CO_331:0001029).

mg : Content of magnesium in dry weight basis measuring mg per 100 g (CO_331:0001030).

bc : Content of beta-carotene in dry weight basis in raw storage roots measuring mg per 100 g (CO_331:0000289).

bc.cc : Content of beta-carotene in fresh weight basis in raw storage roots estimating from RHS color chart in mg per 100 g (CO_331:0001023).

tc : Content of total carotenoids in dry weight basis in raw storage roots measuring mg per 100 g (CO_331:0000290).

star : Content of starch in dry weight basis in raw storage roots measuring percentage (CO_331:0000291).

star.b : Content of starch in dry weight basis in boiled storage roots measuring percentage (CO_331:2000027).

fruc : Content of fructose in dry weight basis in raw storage roots measuring g per 100 g (CO_331:0001045).

fruc.b : Content of fructose in dry weight basis in boiled storage roots measuring g per 100 g (CO_331:0001046).

gluc : Content of glucose in dry weight basis in raw storage roots measuring g per 100 g (CO_331:0001047).

gluc.b : Content of glucose in dry weight basis in boiled storage roots measuring g per 100 g (CO_331:0001048).

sucr : Content of sucrose in dry weight basis in raw storage roots measuring g per 100 g (CO_331:0001049).

sucr.b : Content of sucrose in dry weight basis in boiled storage roots measuring g per 100 g (CO_331:0001050).

malt : Content of maltose in dry weight basis in raw storage roots measuring g per 100 g (CO_331:0001051).

malt.b : Content of maltose in dry weight basis in boiled storage roots measuring g per 100 g (CO_331:0001052).

—————————– Calculated traits —————————-

tnr : Total number of storage roots per plot (CO_331:0000233).

trw : Total storage root weight per plot in kg (CO_331:0000237).

trw.d : Total storage root weight in dry weight basis per plot in kg (CO_331:0000995).

biom : Biomass yield per plot in kg (CO_331:0000683).

biom.d : Biomass yield in dry weight basis per plot in kg (CO_331:0000997).

cytha : Commercial storage root yield in tons per ha (CO_331:0000809).

cytha.aj : Commercial storage root yield adjusted by number of harvested plants per plot in tons per ha (CO_331:0000998).

rytha : Total storage roots yield in tons per ha (CO_331:0000296).

rytha.aj : Total storage roots yield adjusted by number of harvested plants per plot in tons per ha (CO_331:0000999).

dmry : Storage root dry matter yield in tons per ha (CO_331:0001000).

dmry.aj : Storage root dry matter yield adjusted by number of harvested plants per plot in tons per ha (CO_331:0001001).

vw.d : Dry weight of vines per plot in kg (CO_331:0000996).

fytha : Total foliage yield in tons per ha (CO_331:0000684).

fytha.aj : Total foliage yield adjusted by number of harvested plants per plot in tons per ha (CO_331:0001002).

dmvy : Vine dry matter yield in tons per ha (CO_331:0001005).

dmvy.aj : Vine dry matter yield adjusted by number of harvested plants per plot in tons per ha (CO_331:0001006).

bytha : Biomass yield in tons per ha (CO_331:0001003).

bytha.aj : Biomass yield adjusted by number of harvested plants per plot in tons per ha (CO_331:0001004).

dmby : Biomass dry matter yield in tons per ha (CO_331:0001007).

dmby.aj : Biomass dry matter yield adjusted by number of harvested plants per plot in tons per ha (CO_331:0001008).

acrw : Average commercial storage root weight in kg (CO_331:0000680).

ancrw : Average non-commercial storage root weight in kg (CO_331:0002039).

atrw : Average total storage root weight in kg (CO_331:0002040).

nrpp : Number of storage roots per plant harvested (CO_331:0000230).

nrpsp : Number of storage roots per plant planted (CO_331:0000994).

ncrpp : Number of commercial storage roots per plant harvested (CO_331:0006024).

ncrpsp : Number of commercial storage roots per plant planted (CO_331:0000993).

ypp : Storage roots yield per plant harvested in kg (CO_331:0000681).

ypsp : Storage roots yield per plant planted in kg (CO_331:0000989).

vpp : Vine weight per plant harvested in kg (CO_331:0000990).

vpsp : Vine weight per sowed plant in kg (CO_331:0000991).

rtyldpct : Yield as percent of check (CO_331:0000792).

ci : Commercial index in percentage (CO_331:0000682).

hi : Harvest index in percentage (CO_331:0000302).

shi : Survival index in percentage (CO_331:0000301).

rfr : Root foliage ratio in percentage (CO_331:0001015).

Value

It returns:

The fieldbook data frame with all column names in lowercase and with some possible modifications in the names.

A list of warnings for all the column names that have been changed.

A list of warnings for all the column names not recognized.

Author(s)

Raul Eyzaguirre.

Examples

check.names.sp(pjpz09)

[Package st4gi version 3.1.2 Index]