check.pos {st4gi}	R Documentation

Check row and column positions

Description

This function checks that there is only one genotype in each row and column position. This is a wrapper for ck.pos function.

Usage

check.pos(row, col, rep, dfr)

Arguments

row
The name of the column that identifies the rows.

col
The name of the column that identifies the columns.

rep
The name of the column that identifies the replications.

dfr
The name of the data frame.

Value

For each replication a list of plots (unique row and column position) with more than one genotype.

Author(s)

Raul Eyzaguirre.

Examples

# Create a design
dfr <- cr.rcbd(1:20, 3, 10)
dfr <- dfr$book
# Check positions
check.pos('row', 'col', 'block', dfr)

[Package st4gi version 3.1.2 Index]