ck.abd {st4gi}	R Documentation

Check data for an ABD

Description

This function checks the frequencies of genotypes in an ABD.

Usage

ck.abd(trait, geno, rep, dfr)

Arguments

trait
The name of the column for the trait to analyze.

geno
The name of the column that identifies the genotypes including checks.

rep
The name of the column that identifies the replications.

dfr
The name of the data frame.

Value

The number of checks ng.ck, the number of no checks ng, the number of missing values for checks nmis.ck), the number of missing values for no checks nmis, the number nck.0 and list ck.0 of checks without data, the number nck.1 and list ck.1 of checks with only one datum, the number of checks with at least two data nck.2, the number of checks that appear more than once in a given replication (nck.mult), the number of replications nrep, and the number of rows in the data frame with missing values for factors (nmis.fac).

Author(s)

Raul Eyzaguirre.

Examples

# Create design
dfr <- cr.abd(1:50, c('a', 'b', 'd'), 5, 10)
dfr <- dfr$book
# Create some random data
dfr$y <- rnorm(65)
# Delete some values
dfr[c(1, 5, 7, 56), 'y'] <- NA
# Delete some values for classification factors
dfr[64, 'geno'] <- NA
# Check the design
ck.abd('y', 'geno', 'block', dfr)

[Package st4gi version 3.1.2 Index]