ck.lxt {st4gi}	R Documentation

Check data for a line by tester genetic design

Description

This function checks the data structure for a line by tester genetic design.

Usage

ck.lxt(line, tester, dfr)

Arguments

line
The name of the column that identifies the lines.

tester
The name of the column that identifies the testers.

dfr
The name of the data frame.

Value

Three control values, the number of lines (nlin), and the number of testers (ntes). The control values are:

c1: TRUE if all lines appear as parents.

c2: TRUE if all testers appear as parents.

c3: TRUE if all lines x testers are in the crosses.

Author(s)

Raul Eyzaguirre.

Examples

ck.lxt("line", "tester", lxt)

[Package st4gi version 3.1.2 Index]