ck.rcbd {st4gi}	R Documentation

Check data for a RCBD

Description

This function checks the frequencies of genotypes in a RCBD.

Usage

ck.rcbd(trait, geno, rep, dfr)

Arguments

trait
The name of the column for the trait to analyze.

geno
The name of the column that identifies the genotypes.

rep
The name of the column that identifies the replications.

dfr
The name of the data frame.

Value

The number of genotypes without data (ng.0), the number of genotypes with more than one plot in a given block (ng.2), the number of missing values nmis, the proportion of missing values (pmis), the number of genotypes (ng), the number of replications (nrep), a table with valid cases for each genotype (tf), a table with valid cases for each genotype in each replication (tfr), and the number of rows in the data frame with missing values for factors (nmis.fac).

Author(s)

Raul Eyzaguirre.

Examples

# Create design
dfr <- cr.rcbd(1:20, 3, 10)
dfr <- dfr$book
# Create some random data
dfr$y <- rnorm(60)
# Delete some values
dfr[c(1, 5, 16, 17), 'y'] <- NA
# Check the design
ck.rcbd('y', 'geno', 'block', dfr)

[Package st4gi version 3.1.2 Index]