ck.w {st4gi}	R Documentation

Check data for a Wescott layout

Description

This function checks the grid of checks on the Wescott layout and the number of missing values.

Usage

ck.w(trait, geno, ck1, ck2, row, col, ncb, dfr)

Arguments

trait
The name of the column for the trait to analyze.

geno
The name of the column that identifies the genotypes.

ck1
Name of check 1.

ck2
Name of check 2.

row
The name of the column that identifies the rows.

col
The name of the column that identifies the columns.

ncb
Number of columns between two check columns.

dfr
The name of the data frame.

Value

Four control values (c1, c2, c3, and c4, for the grid of checks, the number of missing values for checks (nmis.ck) and genotypes nmis, the proportion of missing values for checks (pmis.ck) and genotypes (pmis), and the number of rows in the data frame with missing values for factors (nmis.fac).

Author(s)

Raul Eyzaguirre.

Examples

# Create a design
dfr <- cr.w(1:1000, "A", "B", 50, 10)
dfr <- dfr$book
# Create some random data
dfr$y <- rnorm(1125)
# Delete some values
dfr[c(11, 165, 569, 914), 'y'] <- NA
# Check the design
ck.w("y", "geno", "A", "B", "row", "col", 10, dfr)

[Package st4gi version 3.1.2 Index]