clean.data.pt {st4gi}	R Documentation

Clean potato data

Description

This is a wrapper for functions setna.pt and setzero.pt and applies both in that order.

Usage

clean.data.pt(dfr, f = 10)

Arguments

dfr
The name of the data frame.

f
Factor for extreme values detection.

Details

The data frame must use the labels (lower or upper case) listed in function check.names.pt. Then functions setna.pt and setzero.pt are applied to the data.

Value

It returns the data frame with all impossible values set to NA, some values set to 0 and a list of warnings with all the rows that have been modified.

Author(s)

Raul Eyzaguirre.

Examples

dfr <- data.frame(mtwp = c(2.2, 5.0, 3.6, 12, 1600, -4, 0),
                  dm = c(21, 23, 105, 24, -3, 30, NA),
                  nmtp = c(1.3, 10, 11, NA, 2, 5, NA))
clean.data.pt(dfr)

[Package st4gi version 3.1.2 Index]