conplot {st4gi}	R Documentation

Plot connectivity among trials

Description

Function to plot number of genotypes in common among trials.

Usage

conplot(trait, geno, trial, dfr)

Arguments

trait
The name of the column for the trait to plot.

geno
The name of the column that identifies the genotypes.

trial
The name of the column that identifies the trials.

dfr
The name of the data frame.

Value

It returns a plot with the number or percentage of genotypes in common among trials.

Author(s)

Raul Eyzaguirre

Examples

conplot("ttwp", "geno", "trial.name", ptfs)

[Package st4gi version 3.1.2 Index]