convert.co.pt {st4gi}	R Documentation

Potato CO numbers

Description

Interchanges potato short labels with Crop Ontology variable numbers.

Usage

convert.co.pt(dfr, direction = c("labels.to.co", "co.to.labels"))

Arguments

dfr
The name of the data frame.

direction
labels.to.co or co.to.labels

Details

All labels (lower or upper case) listed in function check.names.pt are interchanged with CO_330 variable numbers.

Value

It returns a data frame with all short labels for traits interchanged with Crop Ontology variable numbers.

Author(s)

Raul Eyzaguirre.

Examples

convert.co.pt(potatoyield)

[Package st4gi version 3.1.2 Index]