cr.a01d {st4gi}	R Documentation

Alpha (0,1) design

Description

This function creates the fieldbook and fieldplan for an alpha (0,1) design.

Usage

cr.a01d(
  geno,
  check = NULL,
  nb,
  k,
  nc = NULL,
  breakib = c("no", "yes"),
  serpentine = c("yes", "no"),
  alongside = c("no", "rows", "columns")
)

Arguments

geno
The list of genotypes.

check
The list of checks to repeat in each incomplete block (optional).

nb
Number of complete blocks.

k
Size for the incomplete blocks.

nc
Number of available columns on the field for each complete block .

breakib
If incomplete blocks should be broken in more than one row, default "no".

serpentine
"yes" or "no", default "yes".

alongside
"no" for independent blocks, or "rows" or "columns" if blocks are together alongside rows or columns.

Details

The genotypes are randomly allocated on a field following an alpha (0,1) design. In this design each block is a complete replication that is divided into s incomplete blocks of size k. For any pair of genotypes, the number of times they concur into the same incomplete block is 0 or 1 (hence the name of the design). There are 4 options for this design:

nb = 2 and k <= s.

nb = 3, s odd, and k <= s.

nb = 3, s even, and k <= s-1.

nb = 4, s even and not a multiple of 3, and k <= s.

The blocks and incomplete blocks are disposed alongside the rows.

Value

It returns the fieldbook and fieldplan.

Author(s)

Raul Eyzaguirre.

Examples

cr.a01d(1:100, NULL, 2, 5)
cr.a01d(1:100, NULL, 3, 5, 28)

[Package st4gi version 3.1.2 Index]