cr.rcbd {st4gi}	R Documentation

Randomized Complete Block Design

Description

This function creates the fieldbook and fieldplan for a RCBD.

Usage

cr.rcbd(
  geno,
  nb,
  nc = NULL,
  serpentine = c("yes", "no"),
  alongside = c("no", "rows", "columns")
)

Arguments

geno
The list of genotypes.

nb
Number of blocks.

nc
Number of available columns on the field for each block.

serpentine
"yes" or "no", default "yes".

alongside
"no" for independent blocks, or "rows" or "columns" if blocks are together alongside rows or columns.

Value

It returns the fieldbook and fieldplan.

Author(s)

Raul Eyzaguirre.

Examples

cr.rcbd(1:20, 3, 10)
cr.rcbd(1:20, 2, 7)

[Package st4gi version 3.1.2 Index]