cr.spld {st4gi}	R Documentation

Split-Plot Design

Description

This function creates the fieldbook and fieldplan for a Split-Plot or a Split-Split-Plot design with a RCBD for the whole plots.

Usage

cr.spld(fnames, flevels, nb, nc = NULL, serpentine = c("yes", "no"))

Arguments

fnames
Factors' names for two or three factors. See details.

flevels
A list with the factors' levels.

nb
Number of blocks for the first factor.

nc
Number of columns. See details.

serpentine
"yes" or "no", default "yes".

Details

Two or three factors must be included. The first factor goes to the whole plots, the second to the sub-plots and the third, if any, to the sub-sub-plots. The number of columns "nc" is the maximum available number of columns on the field for the sub-plots or the sub-sub-plots if a third factor is provided. Default is the number of levels of the last factor.

Value

It returns the fieldbook and fieldplan.

Author(s)

Raul Eyzaguirre.

Examples

A <- paste0("a", 1:4)
B <- paste0("b", 1:3)
cr.spld(c("A", "B"), list(A, B), 3)

[Package st4gi version 3.1.2 Index]