cr.w {st4gi}	R Documentation

Unreplicated experiment with a grid of checks

Description

This function creates the fieldbook and fieldplan for an unreplicated experiment with genotypes randomly allocated on a field with checks following the method described by Westcott (1981).

Usage

cr.w(geno, ck1, ck2, nc = NULL, ncb = 10, serpentine = c("yes", "no"))

Arguments

geno
The list of genotypes.

ck1
Name of check 1.

ck2
Name of check 2.

nc
Number of available columns on the field.

ncb
Number of columns between two check columns (default is 10).

serpentine
"yes" or "no", default "yes".

Details

The genotypes are randomly allocated on a field between equally spaced columns of two alternating check varieties. Check columns are planted each ncb columns.

Value

It returns the fieldbook and fieldplan.

Author(s)

Raul Eyzaguirre.

References

Westcott, B. (1981). Two methods for early generation yield assessment in winter wheat. In: Proc. of the 4th meeting of the Biometrics in Plant Breeding Section of Eucarpia. INRA Poitier, France, pp 91-95.

Examples

cr.w(1:100, "A", "B", 23)

[Package st4gi version 3.1.2 Index]