create.fb.sp {st4gi}	R Documentation

Create sweetpotato fieldbook

Description

Creates a sweetpotato fieldbook with short labels.

Usage

create.fb.sp(
  design,
  label = c("standard", "CO"),
  minimal = TRUE,
  add = NULL,
  computation = FALSE
)

Arguments

design
The name of the design data frame.

label
Use standard for standard sweetpotato short labels or CO for CO numbers. Default is standard.

minimal
Logical, if TRUE, a minimal list of traits is included. Default is TRUE.

add
Additional traits to include. Only if minimal = TRUE.

computation
Logical, if TRUE, computed traits are included. Only if minimal = FALSE. Default is FALSE.

Details

Only labels listed in function check.names.sp are valid. Uppercase labels are converted to lowercase.

Value

It returns a data frame with fieldbook design and traits.

Author(s)

Raul Eyzaguirre.

Examples

book <- cr.rcbd(1:20, 3, 10)$book
# Get fieldbook with minimal set of traits
create.fb.sp(book)
# Add additional traits
create.fb.sp(book, add = c('bc', 'fe', 'zn'))
# Get the fieldbook with CO numbers
create.fb.sp(book, label = 'CO', add = c('bc', 'fe', 'zn'))

[Package st4gi version 3.1.2 Index]