create.md {st4gi}	R Documentation

Create metadata template file

Description

Creates metadata template file for potatobase and sweetpotatobase upload.

Usage

create.md(
  folder_name = "folder_name",
  trial_name = NA,
  breeding_program = NA,
  location = NA,
  year = NA,
  transplanting_date = NA,
  design_type = NA,
  description = NA,
  trial_type = NA,
  plot_width = NA,
  plot_length = NA,
  field_size = NA,
  planting_date = NA,
  harvest_date = NA,
  number_of_plants_per_ridge = NA,
  number_of_ridges_per_plot = NA,
  space_between_ridges = NA,
  space_between_plants_in_ridge = NA,
  number_plants_per_plot = NA
)

Arguments

folder_name
Name of the folder (optional).

trial_name
Trial name following convention.

breeding_program
Breeding program.

location
Location.

year
Year (format YYYY).

transplanting_date
Transplanting_date (optional with format YYYY-MM-DD),

design_type
Short hand design type. See details.

description
Description of the trial.

trial_type
Trial type. See details.

plot_width
The width of the plot in m (optional).

plot_length
The length of the plot in m (optional).

field_size
The size of the field in ha (optional).

planting_date
Planting date (format YYYY-MM-DD).

harvest_date
Harvest date (format YYYY-MM-DD).

number_of_plants_per_ridge
Required to compute tons/ha (optional).

number_of_ridges_per_plot
Required to compute tons/ha (optional).

space_between_ridges
Required to compute tons/ha.

space_between_plants_in_ridge
Required to compute tons/ha.

number_plants_per_plot
Required to compute tons/ha.

Details

Available designs are:

CRD: for Completely Randomized design

RCBD: for Randomized Complete Block design

Alpha: for Alpha Lattice design

Augmented: for Augmented design

MAD: for Modified Augmented design

Westcott: for Westcott design

Lattice: for Lattice design

RRC: for resolvable row-column design

p-rep: for partially replicated design.

Available trial types are:

Seedling Nursery

phenotyping_trial

Advanced Yield Trial

Preliminary Yield Trial

Uniform Yield Trial

Variety Release Trial

Lattice: for Lattice design

Clonal Evaluation

genetic_gain_trial

storage_trial

heterosis_trial

health_status_trial

grafting_trial

crossing_block_trial

Specialty Trial

Seed Multiplication

Screen House

Fodder trial

Sensory Trial

genotyping_project

Value

It returns a data.frame with metadata to be used with function create.dd.

Author(s)

Raul Eyzaguirre.

Examples

# Empty metadata file
metadata <- create.md()
# Not empty metadata file
trials <- c('PEP2021BAR-AT03', 'PEP2021CAN-AT03', 'PEP2021PIU-AT03',
            'PEP2021SAT-AT03', 'PEP2021PUC-AT03', 'PEP2021CSR-AT03')
locations <- c('Barranca', 'Cañete', 'Piura', 'Satipo', 'Pucallpa',
               'CIP San Ramon Experimental Station')
description <- 'Number of clones: 35 + 5 checks. Number of reps: 2'
pd <- c('2021-11-13', '2021-08-14', '2021-11-16',
        '2021-12-21', '2021-12-21', '2021-12-21')
hd <- c('2022-03-12', '2021-12-14', '2022-03-18',
        '2022-04-22', '2022-04-19', '2022-04-20')                          
metadata <- create.md(trial_name = trials,
                      breeding_program = 'Peru-CIP',
                      location = locations,
                      year = 2021,
                      design_type = 'RRC',
                      description = description,
                      trial_type = 'Advanced Yield Trial',
                      plot_width = 3.6,
                      plot_length = 5.25,
                      planting_date = pd,
                      harvest_date = hd,
                      number_of_plants_per_ridge = 20,
                      number_of_ridges_per_plot = 4,
                      space_between_ridges = 0.9,
                      space_between_plants_in_ridge = 0.25,
                      number_plants_per_plot = 80)

[Package st4gi version 3.1.2 Index]