docomp {st4gi}	R Documentation

Do computations over some factors

Description

Do computations for several traits for some specific factors.

Usage

docomp(do, traits, factors, keep = NULL, dfr, method = c("fast", "slow"))

Arguments

do
The computation to perform. Implemented options are count, mode, and standard functions like mean, median, min, max, sd, var, sum, etc.

traits
The names of the columns for the traits.

factors
The names of the columns for the factors.

keep
The names of additional columns to keep.

dfr
The name of the data frame.

method
Use fast or slow method.

Details

This function do a specific computation for all the traits for each level's combination of the factors. Additional columns can be kept if specified in keep. All factors and keep values are converted to character. do = "count" counts the number of valid cases (excluding missing values).

Value

It returns a data frame with the computations.

Author(s)

Raul Eyzaguirre.

Examples

# Compute means across replications and then across locations for each genotype
traits <- c("rytha", "bc", "dm", "star", "nocr")
factors <- c("geno", "loc")
output1 <- docomp("mean", traits, factors, dfr = spg)
docomp("mean", traits, "geno", dfr = output1) 
# Compute maxima across replications for each genotype and location.
docomp("max", traits, factors, dfr = spg)

[Package st4gi version 3.1.2 Index]