elston {st4gi}	R Documentation

Elston Index

Description

Function to compute the Elston index (Elston, R. C., 1963).

Usage

elston(
  traits,
  geno,
  env = NULL,
  rep = NULL,
  dfr,
  means = c("single", "fitted"),
  lb = 1
)

Arguments

traits
The names of the columns for the traits.

geno
The name of the column that identifies the genotypes.

env
The name of the column that identifies the environments.

rep
The name of the column that identifies the replications.

dfr
The name of the data frame.

means
The genotypic means to compute the index, "single" or "fitted". The default is "single". See details for more information.

lb
Lower bound. 1 for k = min(x) and 2 for k = (n \times min(x) - max(x)) / (n - 1)

Details

The Elston index is a weight free index. It is assumed that all the traits are in the same direction where the highest the value the better the genotype. To include any trait with an opposite direction it must be transformed by multiplication by -1 before.

If means = "fitted" then a linear model is fitted and used to estimate the means for each genotype. If means = "single" and env is specified, then single arithmetic means are computed over the replications for each genotype at each environment and then for each genotype over environments. If means = "single" and env is not specified, then single arithmetic means are computed over all the observations for each genotype.

Value

It returns a data frame with the genotypic means for each trait, the Elston index, and the rank for each genotype according to the index.

Author(s)

Raul Eyzaguirre.

References

Elston, R. C. (1963). A weight-free index for the purpose of ranking or selection with respect to several traits at a time. Biometrics. 19(1): 85-97.

Examples

elston(c("rytha", "bc", "dm", "star", "nocr"), "geno", dfr = spg)

[Package st4gi version 3.1.2 Index]