mve.met {st4gi}	R Documentation

Estimation of missing values for a MET in a RCBD

Description

Function to estimate missing values for a Multi Environment Trial (MET) with a Randomized Complete Block Design (RCBD) by the least squares method.

Usage

mve.met(trait, geno, env, rep, dfr, maxp = 0.1, tol = 1e-06)

Arguments

trait
The name of the column for the trait to estimate missing values.

geno
The name of the column that identifies the genotypes.

env
The name of the column that identifies the environments.

rep
The name of the column that identifies the replications.

dfr
The name of the data frame.

maxp
Maximum allowed proportion of missing values to estimate, default is 10%.

tol
Tolerance for the convergence of the iterative estimation process.

Value

It returns a data frame with the experimental layout and columns trait and trait.est with the original data and the original data plus the estimated values.

Author(s)

Raul Eyzaguirre.

Examples

mve.met("y", "geno", "env", "rep", met8x12)

[Package st4gi version 3.1.2 Index]