optim.w {st4gi}	R Documentation

Optimize Westcott adjustment

Description

This function optimizes the parameter values of the Westcott adjustment.

Usage

optim.w(
  trait,
  geno,
  replicated,
  ck1,
  ck2,
  row,
  col,
  ncb,
  nrs,
  method,
  ind,
  p,
  opt = c("F.value", "MSE", "CV"),
  dfr
)

Arguments

trait
The name of the column for the trait to adjust.

geno
The name of the column that identifies the genotypes.

replicated
A yes/no column to identify replicated genotypes.

ck1
Name of check 1.

ck2
Name of check 2.

row
The name of the column that identifies the rows.

col
The name of the column that identifies the columns.

ncb
Number of columns between two check columns.

nrs
Set of values for the nrs param on aj.w.

method
Set of values for the method param on aj.w.

ind
Set of values for the ind param on aj.w.

p
Set of values for the p param on aj.w.

opt
Criteria for optimization, F.value, MSE or CV. See details.

dfr
The name of the data frame.

Details

If optim = F.value then the F value for ANOVA is maximized. If optim = MSE then the MSE is minimized. If optim = CV then the CV is minimized.

Value

It returns optimal set of values under the selected criteria.

Author(s)

Raul Eyzaguirre.

[Package st4gi version 3.1.2 Index]