pesekbaker {st4gi}	R Documentation

Pesek-Baker Index

Description

Function to compute the Pesek-Baker index (Pesek, J. and R.J. Baker., 1969).

Usage

pesekbaker(
  traits,
  geno,
  env = NULL,
  rep,
  dfr,
  gvals = c("blups", "blues", "means"),
  dgg = NULL,
  units = c("sdu", "actual"),
  sf = 0.1,
  method = 1
)

Arguments

traits
The names of the columns for the traits.

geno
The name of the column that identifies the genotypes.

env
The name of the column that identifies the environments.

rep
The name of the column that identifies the replications.

dfr
The name of the data frame.

gvals
The genotypic values to compute the index, "blups", "blues" or "means". The default is "blups".

dgg
Desired genetic gains. The default is one standard deviation for each trait.

units
Units for dgg, "sdu" or "actual". See details for more information.

sf
Selected fraction. The default is 0.1.

method
The method to compute genotypic covariances. See ?ecm for details.

Details

The Pesek-Baker is an index where relative economic weights have been replaced by desired gains.

If dgg is not specified, the desired genetic gains are set to one standard deviation for each trait. dgg can be specified in actual units (units = "actual") or in standard deviations (units = "sdu"), defaults to "sdu". For example, if you have a trait which is expressed in kilograms and with a standard deviation of 5 kilograms, typing dgg = 2 means a desired genetic gain of 2 standard deviations that corresponds to 10 kilograms. If you type dgg = 2 and units = "actual" then this means a desired genetic gain of 2 kilograms. If dgg = NULL then the desired genetic gains will be one standard deviation, no matter if units is set as "actual" or "sdu".

Value

It returns:

$Desired.Genetic.Gains, the desired genetic gains in actual units,

$Standard.Deviations, the estimated genotypic standard deviations,

$Index.Coefficients, the index coefficients,

$Response.to.Selection, the response to selection,

$Std.Response.to.Selection, the standardized response to selection, and

$Pesek.Baker.Index, a data frame with the genotypic means for each trait, the Pesek-Baker index, and the rank for each genotype according to the index.

Author(s)

Raul Eyzaguirre.

References

Pesek, J. and R.J. Baker.(1969). Desired improvement in relation to selection indices. Can. J. Plant. Sci. 9:803-804.

Examples

traits <- c("rytha", "bc", "dm", "star", "nocr")
pesekbaker(traits, "geno", "loc", "rep", spg)
# More weight on bc and dm, less on star and nocr
pesekbaker(traits, "geno", "loc", "rep", spg, dgg = c(1, 1.5, 1.5, 0.8, 0.8))

[Package st4gi version 3.1.2 Index]