plot.st4gi_ammi {st4gi}	R Documentation

AMMI or GGE biplots

Description

This function produces AMMI (Gollob, H. R., 1968) or GGE (Yan , W. et al., 2000) biplots using the base system and ggplot2.

Usage

## S3 method for class 'st4gi_ammi'
plot(
  x,
  bp.type = 2,
  bp1.type = c("effects", "means"),
  graph.type = c("base", "ggplot"),
  color = c("darkorange", "black", "gray"),
  main = NULL,
  xlab = NULL,
  ...
)

Arguments

x
An object of class ammi.

bp.type
Choose 1 for the trait-PC1 biplot and 2 for the PC1-PC2 biplot.

bp1.type
Choose "effects" or "means" for biplot-1.

graph.type
"base" or "ggplot".

color
Color for lines, symbols and/or labels for environments, genotypes and axes (Only for the base system plot).

main
Main title.

xlab
A title for the x axis.

...
Additional plot arguments (Only for the base system plot).

Details

It produces a biplot for an object of class ammi. See ?ammi for additional details.

Value

It returns a dispersion plot of means or effects against the first PC, or a dispersion plot of PC1 against PC2.

Author(s)

Raul Eyzaguirre.

References

Gollob, H. R. (1968). A Statistical Model which combines Features of Factor Analytic and Analysis of Variance Techniques, Psychometrika, Vol 33(1): 73-114.

Yan, W. et al. (2000). Cultivar evaluation and mega-environment investigation based on the GGE biplot, Crop Sci., Vol 40: 597-605.

Examples

model.ammi <- ammi("y", "geno", "env", "rep", met8x12)
plot(model.ammi)
plot(model.ammi, bp.type = 1)
plot(model.ammi, graph.type = "ggplot")

[Package st4gi version 3.1.2 Index]