plot.st4gi_tai {st4gi}	R Documentation

Tai plot

Description

This function produces a Tai's stability analysis plot (Tai, G. C. C., 1971) using the base system and ggplot2.

Usage

## S3 method for class 'st4gi_tai'
plot(
  x,
  conf = 0.95,
  graph.type = c("base", "ggplot"),
  color = c("darkorange", "black", "gray"),
  ...
)

Arguments

x
An object of class tai.

conf
Probability for the Tai limits.

graph.type
"base" or "ggplot".

color
Color for symbols, labels and lines (Only for the base system plot).

...
Additional plot arguments (Only for the base system plot).

Details

The limits for alpha and lambda are computed using the mean squares from an ANOVA table for a RCBD with blocks nested into environments.See ?tai for additional details.

Value

It returns the Tai graph for stability analysis.

Author(s)

Raul Eyzaguirre.

References

Tai, G. C. C. (1971). Genotypic Stability Analysis and Its Application to Potato Regional Trials, Crop Science, Vol 11.

Examples

model.tai <- tai("y", "geno", "env", "rep", met8x12)
plot(model.tai)
plot(model.tai, graph.type = "ggplot")

[Package st4gi version 3.1.2 Index]