rsa {st4gi}	R Documentation

Regression Stability Analysis

Description

Function to run the regression stability analysis (Yates and Cochran, 1938, Finlay and Wilkinson, 1963).

Usage

rsa(trait, geno, env, rep, dfr, maxp = 0.1)

Arguments

trait
The name of the column for the trait to analyze.

geno
The name of the column that identifies the genotypes.

env
The name of the column that identifies the environments.

rep
The name of the column that identifies the replications.

dfr
The name of the data frame.

maxp
Maximum allowed proportion of missing values to estimate, default is 10%.

Details

The regression stability analysis is evaluated with a balanced data set. If data is unbalanced, missing values are estimated up to an specified maximum proportion, 10% by default. For the ANOVA table, genotypes and environments are considered as fixed factors while the blocks are considered as random and nested into the environments. To run a regression stability analysis you need a set of genotypes evaluated in a set of environments. At least 3 genotypes or environments are needed. In a regression stability analysis for genotypes grown at several environments, for each genotype a simple linear regression of individual yield (Y) on the mean yield of all genotypes for each environment (X) is fitted. In a similar way, for each environment a simple linear regression of individual yield (Y) on the mean yield of all environments for each genotype (X) is fitted. In both cases the X values are centered on zero, so the intercepts of the models correspond to the means of the genotypes or environments.

Value

It returns the regression stability analysis decomposition of the GxE interaction for genotypes and environments (Heterogeneity among regressions and deviation from regression), the coefficient of variation, and the following regression stability measures for genotypes and environments:

a the intercept.

b the slope.

se the standard error for the slope.

MSe the mean square error.

MSentry the variance of the genotype means across environments and the environment means across genotypes.

MSinter the variance of the genotype interaction effects across environments and the environment interaction effects across genotypes.

Author(s)

Raul Eyzaguirre.

References

Finlay, K. W., and Wilkinson, G. N. (1963). The Analysis of Adaption in a Plant-Breeding Programme. Aust. J. Agric. Res. 14: 742-754.

Yates, F., and Cochran, W. G. (1938). The Analysis of Group Experiments. J. Agric. Sci. 28: 556-580.

Examples

rsa("y", "geno", "env", "rep", met8x12)

[Package st4gi version 3.1.2 Index]