rts1 {st4gi}	R Documentation

Response to selection for a single experiment

Description

It finds the optimum number of replications to get the maximum response to selection for a single experiment for a given plot capacity, number of selected genotypes, genotypic variance and error variance.

Usage

rts1()

Details

It uses package shiny for the web layout. Type rts1() in the R console to run the app.

Value

It returns a plot of response to seleccion versus number of replications and computes the optimum number of replications and the response to selection at this optimum value.

Author(s)

Raul Eyzaguirre.

[Package st4gi version 3.1.2 Index]