rts4 {st4gi}	R Documentation

Response to selection with several locations in two steps (two years)

Description

It computes the response to selection for each step in a two steps selection with several locations for a given number of genotypes at step 1, the number of locations, replications and selected genotypes at step 1, the number of locations, replications and selected genotypes at step 2, and the genotypic, genotypic by location, genotypic by year, genotypic by location by year, and error variances.

Usage

rts4()

Details

It uses package shiny for the web layout. Type rts4() in the R console to run the app.

Value

It returns the response to selection at step 1 and 2.

Author(s)

Raul Eyzaguirre.

[Package st4gi version 3.1.2 Index]