sti {st4gi}	R Documentation

Compute stress tolerance indices

Description

Compute several stress tolerance indices.

Usage

sti(traits, geno, normal, stress, dfr)

Arguments

traits
The list of traits.

geno
The name of the column that identifies the genotypes.

normal
The identification for the normal group.

stress
The identification for the stress group.

dfr
The name of the data frame.

Details

The names of the columns for the traits must follow the convention trait.normal and trait.stress, so the data frame must be in wide format for the factor with levels normal and stress.

For a trait y with values y_n and y_s under normal and stress conditions, the stress intensity (si) is computed:

si = 1 - \frac{\bar y_s}{\bar y_n}

It ranges between 0 and 1 and the larger the value, the more severe is the stress intensity. Then, the following indices are computed:

tol : Tolerance.

tol = y_n - y_s

yrr : Yield reduction ratio.

yrr = 1 - \frac{y_s}{y_n}

ssi : Stress susceptibility index.

ssi = \frac{yrr}{si}

sti : Stress tolerance index.

sti = \frac{y_n \times y_s}{\bar y_n^2}

mp : Mean productivity.

mp = \frac{y_n + y_s}{2}

gmp : Geometric mean productivity.

gmp = √{y_n \times y_s}

Value

It returns a data frame with the indices and a data frame with the stress intensity values for each trait. The names for the indices follow the convention trait.index.

Author(s)

Raul Eyzaguirre.

References

Fernandez, G.C.J. (1992). Effective Selection Criteria for Assessing Stress Tolerance. In: Kuo, C.G., Ed., Proceedings of the International Symposium on Adaptation of Vegetables and Other Food Crops in Temperature and Water Stress, AVRDC Publication, Tainan, 257-270.

Examples

traits <- c("nmtp", "mtwp", "nnomtp")
sti(traits, 'genotype', 'DTWW', 'DTWS', potatostress)

[Package st4gi version 3.1.2 Index]