tai {st4gi}	R Documentation

Tai's stability analysis

Description

This function runs Tai's stability analysis (Tai, G. C. C., 1971). It assumes a RCBD with fixed effects for genotypes and random effects for environments.

Usage

tai(trait, geno, env, rep, dfr, maxp = 0.1)

Arguments

trait
The name of the column for the trait to analyze.

geno
The name of the column that identifies the genotypes.

env
The name of the column that identifies the environments.

rep
The name of the column that identifies the replications.

dfr
The name of the data frame.

maxp
Maximum allowed proportion of missing values to estimate, default is 10%.

Details

If the data set is unbalanced, a warning is produced.

Value

It returns the alpha and lambda values for each genotype for the Tai stability analysis.

Author(s)

Raul Eyzaguirre.

References

Tai, G. C. C. (1971). Genotypic Stability Analysis and Its Application to Potato Regional Trials, Crop Science, Vol 11.

Examples

model.tai <- tai("y", "geno", "env", "rep", met8x12)
model.tai$Tai_values

[Package st4gi version 3.1.2 Index]